pyproject.toml
[tool.poetry]
name = "wallaby"
version = "0.1.0"
description = ""
authors = ["John Casey <jdcasey@commonjava.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"
python-dateutil = "^2.9.0.post0"
pdpyras = "^5.3.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^7.4.0"
pytest-mock = "^3.12.0"
freezegun = "^1.2.2"

[tool.poetry.scripts]
wlb = "wallaby.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
